# [OpenCL] Fuse convolution with ReLU