--- a/lib/Backends/OpenCL/OpenCL.h
+++ b/lib/Backends/OpenCL/OpenCL.h
@@ -244,8 +244,15 @@ public:
 
   bool shouldLower(const Node *N) const override {
     // The group convolution is supported in OpenCL slow convolution kernel.
-    if (N->getKind() == Kinded::Kind::ConvolutionNodeKind)
+    if (N->getKind() == Kinded::Kind::ConvolutionNodeKind) {
       return false;
+    }
+    // Do not lower ReLU to max, but let it pass to the backend where we
+    // can implement it with a unary max(0, x) kernel. This also enables fusing
+    // convolution with ReLU.
+    if (N->getKind() == Kinded::Kind::ReluNodeKind) {
+      return false;
+    }
     return true;
   }
 
@@ -254,6 +261,14 @@ public:
 
   runtime::DeviceManager *
   createDeviceManager(const runtime::DeviceConfig &deviceConfig) override;
+
+  /// \returns whether the backend supports fusing \p activation into \p parent.
+  bool supportsFusedActivation(Node *parent, Node *activation) const override {
+    // Only support convolution+relu fusions for now.
+    bool V = parent->getKind() == Kinded::Kind::ConvolutionNodeKind &&
+             activation->getKind() == Kinded::Kind::ReluNodeKind;
+    return V;
+  }
 
 private:
   /// Parses the graph \F and builds a TraceInfo structure from any found