--- a/lib/Backends/OpenCL/kernels_fwd_quantized_conv.cl
+++ b/lib/Backends/OpenCL/kernels_fwd_quantized_conv.cl
@@ -35,6 +35,7 @@
 // v_imso_0, v_imso_1 - Spacial dimensions of output.
 // v_pad_A - Padding required for tiles of A.
 // v_pad_B - Padding required for tiles of A.
+// v_fuse_relu - Apply max(x, 0) to the output.
 // workgroup_size_0, workgroup_size_1 - workgroup sizes along each dimension.
 // VWN - vector width in dimension N.
 // VWM - vector width in dimension M.
@@ -338,6 +339,10 @@ __kernel
               (float)(((int *)(&(Creg[wm][wn / VWN])))[wn % VWN] + biasval) *
                   (matMulScale / c_scale) +
               c_offset));
+          if (v_fuse_relu) {
+            Cptr[globalRow * NN + globalCol] =
+                max(Cptr[globalRow * NN + globalCol], (Dtype)0);
+          }
         }
       }
     }