* Intel SDK (Intel i7-7500U CPU)

inception_v1: From 0.90 to 0.71 (22% drop)
mobilenetv2:  From 0.40 to 0.16 (60% drop)
resnet50:     From 1.41 to 1.25 (11% drop)

* Intel GPU (HD620)

inception_v1: From 0.14 to 0.14 (0% drop)
mobilenetv2:  From 0.21 to 0.20 (5% drop)
resnet50:     From 0.29 to 0.29 (0% drop)

* pocl / LLVM 9 (Intel i7-7500U CPU)

inception_v1: From 0.80 s to 0.52 s (35% drop)
mobilenetv2:  From 0.23 s to 0.15 s (35% drop)
resnet50:     From 1.97 s to 1.25 s (36% drop)

Test Plan:

Tested&benchmarked with the above OpenCL envs and nets.