@pjaaskel 2020-02-21
> I fixed the typo, but I (still) couldn't reproduce those failures. They seem relevant though, as they concern ReLU. I'll try a clean build.

_ninja clean && ninja && ninja check_  passes here... strange.