@opti-mix 2020-02-21
@pjaaskel The CI error is:
```
./glow/AutoGenInstr.h:2607: void glow::ReluInst::verify() const: Assertion `getDest()->getType() == getSrc()->getType() && "Invalid Type"' failed.
```

It seems like there is a ReluInst with incorrect operand types being produced.