@pjaaskel 2020-02-21
OK, it finally reproduced locally after a complete Glow rebuild from the scratch.

The ReLU in question here:
```
7 %fc_bias_res = allocactivation  { Ty: i8[S:0.1043 O:-104][-2.504,24.101]<2 x 16>} // size: 32 // Users: @out 19, @in 10, @out 8
... 
9 %relu_res = allocactivation  { Ty: i8[S:0.0945 O:-128][0.000,24.109]<2 x 16>} // size: 32 // Users: @out 20, @in 12, @out 10 
10 %relu = relu @out %relu_res, @in %fc_bias_res
```

Thus, the inputs/outputs have different quantization parameters. Perhaps this happens because previously it has been lowered to max which does not check for the type. Just guessing, as I didn't look deeper yet (I have to get back to this tomorrow), perhaps you have educated hints on how to fix this?