@opti-mix 2020-02-22
@pjaaskel OK, all tests pass now after #4205 had landed.

@jfix71 Any further comments? Are you ok with the introduced foldActivations() call?